put shared logic into common component + fixes
fixes: #1398
