+++ src/identity-providers/add/SamlConnectSettings.tsx
@@ -1,26 +1,17 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { useFormContext } from "react-hook-form";
-import {
-  FormGroup,
-  Switch,
-  TextInput,
-  Title,
-  ValidatedOptions,
-} from "@patternfly/react-core";
+import { FormGroup, TextInput, Title } from "@patternfly/react-core";
 
 import { HelpItem } from "../../components/help-enabler/HelpItem";
 import { useTranslation } from "react-i18next";
 import { useAdminClient } from "../../context/auth/AdminClient";
 import type IdentityProviderRepresentation from "@keycloak/keycloak-admin-client/lib/defs/identityProviderRepresentation";
 
-import { JsonFileUpload } from "../../components/json-file-upload/JsonFileUpload";
+import { FileUploadForm } from "../../components/json-file-upload/FileUploadForm";
 import { useRealm } from "../../context/realm-context/RealmContext";
 import { DescriptorSettings } from "./DescriptorSettings";
 import { getBaseUrl } from "../../util";
-
-type Result = IdentityProviderRepresentation & {
-  error: string;
-};
+import { DiscoveryEndpointField } from "../component/DiscoveryEndpointField";
 
 export const SamlConnectSettings = () => {
   const { t } = useTranslation("identity-providers");
@@ -28,16 +19,7 @@ export const SamlConnectSettings = () => {
 
   const adminClient = useAdminClient();
   const { realm } = useRealm();
-  const { setValue, register, errors } = useFormContext();
-
-  const [descriptor, setDescriptor] = useState(true);
-
-  const [entityUrl, setEntityUrl] = useState("");
-  const [descriptorUrl, setDescriptorUrl] = useState("");
-  const [discovering, setDiscovering] = useState(false);
-  const [discoveryResult, setDiscoveryResult] = useState<Result>();
-
-  const defaultEntityUrl = `${getBaseUrl(adminClient)}realms/${realm}`;
+  const { setValue, register, errors, setError } = useFormContext();
 
   const setupForm = (result: IdentityProviderRepresentation) => {
     Object.entries(result).map(([key, value]) =>
@@ -45,38 +27,10 @@ export const SamlConnectSettings = () => {
     );
   };
 
-  useEffect(() => {
-    if (!discovering) {
-      return;
-    }
-
-    setDiscovering(!!entityUrl);
-
-    if (!entityUrl) {
-      return;
-    }
-
-    (async () => {
-      let result;
-      try {
-        result = await adminClient.identityProviders.importFromUrl({
-          providerId: id,
-          fromUrl: entityUrl,
-        });
-      } catch (error) {
-        result = { error };
-      }
-
-      setDiscoveryResult(result as Result);
-      setupForm(result);
-      setDiscovering(false);
-    })();
-  }, [discovering]);
-
-  const fileUpload = async (obj: object) => {
+  const fileUpload = async (xml: string) => {
     const formData = new FormData();
     formData.append("providerId", id);
-    formData.append("file", new Blob([JSON.stringify(obj)]));
+    formData.append("file", new Blob([xml]));
 
     try {
       const response = await fetch(
@@ -93,8 +47,11 @@ export const SamlConnectSettings = () => {
       );
       const result = await response.json();
       setupForm(result);
-    } catch (error: any) {
-      setDiscoveryResult({ error });
+    } catch (error) {
+      setError("discoveryError", {
+        type: "manual",
+        message: (error as Error).message,
+      });
     }
   };
 
@@ -120,90 +77,39 @@ export const SamlConnectSettings = () => {
           name="config.entityId"
           data-testid="serviceProviderEntityId"
           id="kc-service-provider-entity-id"
-          value={entityUrl || defaultEntityUrl}
-          onChange={setEntityUrl}
           ref={register()}
         />
       </FormGroup>
 
-      <FormGroup
-        label={t("useEntityDescriptor")}
-        fieldId="kc-use-entity-descriptor"
-        labelIcon={
-          <HelpItem
-            helpText="identity-providers-help:useEntityDescriptor"
-            forLabel={t("useEntityDescriptor")}
-            forID="kc-use-entity-descriptor-switch"
-          />
+      <DiscoveryEndpointField
+        id="saml"
+        fileUpload={
+          <FormGroup
+            label={t("importConfig")}
+            fieldId="kc-import-config"
+            labelIcon={
+              <HelpItem
+                helpText="identity-providers-help:importConfig"
+                forLabel={t("importConfig")}
+                forID="kc-import-config"
+              />
+            }
+            validated={errors.discoveryError ? "error" : "default"}
+            helperTextInvalid={errors.discoveryError}
+          >
+            <FileUploadForm
+              id="kc-import-config"
+              extension=".xml"
+              hideDefaultPreview
+              unWrap
+              validated={errors.discoveryError ? "error" : "default"}
+              onChange={(value) => fileUpload(value)}
+            />
+          </FormGroup>
         }
       >
-        <Switch
-          id="kc-use-entity-descriptor-switch"
-          label={t("common:on")}
-          data-testid="useEntityDescriptor"
-          labelOff={t("common:off")}
-          isChecked={descriptor}
-          onChange={setDescriptor}
-        />
-      </FormGroup>
-
-      {descriptor && (
-        <FormGroup
-          label={t("samlEntityDescriptor")}
-          fieldId="kc-saml-entity-descriptor"
-          labelIcon={
-            <HelpItem
-              helpText="identity-providers-help:samlEntityDescriptor"
-              forLabel={t("samlEntityDescriptor")}
-              forID="kc-saml-entity-descriptor"
-            />
-          }
-        >
-          <TextInput
-            type="text"
-            name="samlEntityDescriptor"
-            data-testid="samlEntityDescriptor"
-            id="kc-saml-entity-descriptor"
-            value={descriptorUrl}
-            onChange={setDescriptorUrl}
-            ref={register()}
-            validated={
-              errors.samlEntityDescriptor
-                ? ValidatedOptions.error
-                : ValidatedOptions.default
-            }
-          />
-        </FormGroup>
-      )}
-      {!descriptor && (
-        <FormGroup
-          label={t("importConfig")}
-          fieldId="kc-import-config"
-          labelIcon={
-            <HelpItem
-              helpText="identity-providers-help:importConfig"
-              forLabel={t("importConfig")}
-              forID="kc-import-config"
-            />
-          }
-          validated={discoveryResult?.error ? "error" : "default"}
-          helperTextInvalid={discoveryResult?.error.toString()}
-        >
-          <JsonFileUpload
-            id="kc-import-config"
-            helpText="identity-providers-help:jsonFileUpload"
-            hideDefaultPreview
-            unWrap
-            validated={discoveryResult?.error ? "error" : "default"}
-            onChange={(value) => fileUpload(value)}
-          />
-        </FormGroup>
-      )}
-
-      {descriptor && discoveryResult && !discoveryResult.error && (
-        <DescriptorSettings readOnly={true} />
-      )}
-      {!descriptor && <DescriptorSettings readOnly={false} />}
+        {(readonly) => <DescriptorSettings readOnly={readonly} />}
+      </DiscoveryEndpointField>
     </>
   );
 };
